test(loop): split observation contract coverage
Constraint: Keep touched Python test modules below 250 pure LOC with one contract responsibility each.
Rejected: combined phase-history and observation-contract test module | exceeded the size ceiling and mixed independent domains.
Confidence: high
Scope-risk: narrow
Directive: Keep phase-transition and goal-observation regressions in separate focused modules.
Tested: 7105 unittest cases; 15 focused split-module cases; Ruff; compileall; workflow generation; release drift; diff check.
Not-tested: live Hermes host event ingestion; basedpyright diagnostics unavailable because the server is not installed.
